Refactor trigger UI components into dedicated module
