+++ internal-packages/workflow-designer-ui/src/header/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./trigger-input-dialog";
